refactor move analysisManager to store/ directory
