+++ src/pageEditor/panes/insert/exampleStarterBrickConfigs.ts
