+++ src/pageEditor/store/analysisManager.ts
@@ -25,12 +25,12 @@ import ReduxAnalysisManager from "@/analysis/ReduxAnalysisManager";
 import { type UUID } from "@/types/stringTypes";
 import { type TraceRecord } from "@/telemetry/trace";
 import { isAnyOf, type PayloadAction } from "@reduxjs/toolkit";
-import { type RootState } from "./store/editor/pageEditorTypes";
+import { type RootState } from "./editor/pageEditorTypes";
 import { actions as editorActions } from "@/pageEditor/store/editor/editorSlice";
-import runtimeSlice from "./store/runtime/runtimeSlice";
+import runtimeSlice from "./runtime/runtimeSlice";
 import RequestPermissionAnalysis from "@/analysis/analysisVisitors/requestPermissionAnalysis";
 import FormBrickAnalysis from "@/analysis/analysisVisitors/formBrickAnalysis";
-import { selectActiveModComponentTraces } from "./store/runtime/runtimeSelectors";
+import { selectActiveModComponentTraces } from "./runtime/runtimeSelectors";
 import VarAnalysis from "@/analysis/analysisVisitors/varAnalysis/varAnalysis";
 import analysisSlice from "@/analysis/analysisSlice";
 import RegexAnalysis from "@/analysis/analysisVisitors/regexAnalysis";
